REFACTORED: merged changes and renamed module to modulbeauftragter (at the controller-, test- and html-file) because of the projectname which is the same.
